Modified by Ahmed wagdi And Mohamed Reda Elnakeeb 1- Menu and Menu Item List 2- Create Menu and Menu Item 3- Assign Access to menu items for groups 3 - create group and menuItem converters

+++ src/main/java/com/ntgclarity/smartcompound/dataaccess/base/BaseDAO.java
@@ -66,7 +66,7 @@ public abstract class BaseDAO {
 
 		Query query = getCurrentSession().createQuery(
 				"select count(x) from " + cls.getCanonicalName() + " x ");
-		int result = (int) query.uniqueResult();
+		int result = ((Long) query.uniqueResult()).intValue();
 		return result;
 	}
 }
